Separate reader into its own file
