+++ src/protocol.rs
@@ -329,165 +329,6 @@ fn read_message_event(buf: &[u8]) -> Result<(usize, Message<'_>), Error> {
 	))
 }
 
-#[cfg(feature = "reader")]
-#[derive(Default)]
-#[cfg_attr(all(feature = "serde", feature = "alloc"), derive(Serialize, Deserialize))]
-pub struct Reader {
-	header_done: bool,
-	buffer: BytesMut,
-}
-
-#[cfg(feature = "reader")]
-impl Reader {
-	pub fn push(&mut self, other: &[u8]) {
-		self.buffer.extend_from_slice(other);
-	}
-	/// Attempt to read from the reader's internal buffer.
-	/// We further translate the data from [`MessageType`] into an owned [`Message`] for use.
-	///
-	/// # Errors
-	///
-	/// See [`read_message_type`] for failure cases.
-	pub fn try_read(&mut self) -> Result<MessageOwned, Error> {
-		let mut data = self.buffer.split().freeze();
-		let (new_buf, message_type) = read_message_type(&data, self.header_done)
-			.map(|(offset, mt)| (BytesMut::from(&data[offset..]), mt))?;
-
-		let msg = match message_type {
-			MessageType::Version { version } => {
-				self.header_done = true;
-				MessageOwned::Version(version.into_iter().collect())
-			}
-			MessageType::Audio { samples_offset, samples_len } => MessageOwned::Audio(
-				data.split_off(samples_offset - 1).split_to(samples_len),
-			),
-			MessageType::Event { typ, start, end, name_offset, name_len } => {
-				MessageOwned::Event(EventOwned {
-					typ,
-					start,
-					end,
-					name: if name_len == 0 {
-						None
-					} else {
-						// TODO: try to remove this clone!
-						Some(data
-							.split_off(name_offset - 1)
-							.split_to(name_len)
-							.into_iter()
-							.map(char::from)
-							.collect::<String>())
-					},
-				})
-			}
-		};
-
-		self.buffer = new_buf;
-		Ok(msg)
-	}
-}
-
-#[cfg(feature = "reader")]
-#[test]
-fn test_wave_reader() {
-	use alloc::string::ToString;
-
-	use assert_matches::assert_matches;
-	let mut reader = Reader::default();
-	let data: &[u8] = include_bytes!("../test.wav");
-	reader.push(data);
-	assert_eq!(reader.try_read(), Ok(MessageOwned::Version("0.01".to_string())));
-	assert_eq!(
-		reader.try_read(),
-		Ok(MessageOwned::Event(EventOwned {
-			typ: EventType::Sentence,
-			start: 0,
-			end: 0,
-			name: None
-		}))
-	);
-	assert_eq!(
-		reader.try_read(),
-		Ok(MessageOwned::Event(EventOwned {
-			typ: EventType::Word,
-			start: 0,
-			end: 4,
-			name: None
-		}))
-	);
-	for i in 0..4 {
-		assert_matches!(reader.try_read(), Ok(MessageOwned::Audio(_)), "{i}");
-	}
-	let word_is = MessageOwned::Event(EventOwned {
-		typ: EventType::Word,
-		start: 5,
-		end: 7,
-		name: None,
-	});
-	let word_a = MessageOwned::Event(EventOwned {
-		typ: EventType::Word,
-		start: 8,
-		end: 9,
-		name: None,
-	});
-	let word_test = MessageOwned::Event(EventOwned {
-		typ: EventType::Word,
-		start: 10,
-		end: 14,
-		name: None,
-	});
-	let word_using = MessageOwned::Event(EventOwned {
-		typ: EventType::Word,
-		start: 15,
-		end: 20,
-		name: None,
-	});
-	let word_spiel = MessageOwned::Event(EventOwned {
-		typ: EventType::Word,
-		start: 21,
-		end: 26,
-		name: None,
-	});
-	let word_whaha = MessageOwned::Event(EventOwned {
-		typ: EventType::Word,
-		start: 28,
-		end: 35,
-		name: None,
-	});
-	assert_eq!(reader.try_read(), Ok(word_is));
-	for _ in 0..3 {
-		assert_matches!(reader.try_read(), Ok(MessageOwned::Audio(_)));
-	}
-	assert_eq!(reader.try_read(), Ok(word_a));
-	assert_matches!(reader.try_read(), Ok(MessageOwned::Audio(_)));
-	assert_matches!(reader.try_read(), Ok(MessageOwned::Audio(_)));
-	assert_eq!(reader.try_read(), Ok(word_test));
-	for _ in 0..6 {
-		assert_matches!(reader.try_read(), Ok(MessageOwned::Audio(_)));
-	}
-	assert_eq!(reader.try_read(), Ok(word_using));
-	for _ in 0..6 {
-		assert_matches!(reader.try_read(), Ok(MessageOwned::Audio(_)));
-	}
-	assert_eq!(reader.try_read(), Ok(word_spiel));
-	for _ in 0..14 {
-		assert_matches!(reader.try_read(), Ok(MessageOwned::Audio(_)));
-	}
-	assert_eq!(
-		reader.try_read(),
-		Ok(MessageOwned::Event(EventOwned {
-			typ: EventType::Sentence,
-			start: 28,
-			end: 28,
-			name: None
-		}))
-	);
-	assert_eq!(reader.try_read(), Ok(word_whaha));
-	for _ in 0..10 {
-		assert_matches!(reader.try_read(), Ok(MessageOwned::Audio(_)));
-	}
-	assert_eq!(&reader.buffer.freeze().slice(..)[..], &[]);
-}
-
 #[test]
 fn test_read_write_version() {
 	let mt = Message::Version("wowz");
